fixed character creation behavior
